Re-organize components & routes
