+++ src/App.jsx
@@ -3,6 +3,7 @@ import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom';
 import { HomePage } from 'pages/HomePage/HomePage';
 import { ReplyListPage } from 'pages/ReplyListPage/ReplyListPage';
 import { UserPage } from 'pages/UserPage/UserPage';
+import { FollowPage } from 'pages/FollowPage/FollowPage';
 import { AdminLoginPage } from 'pages/AdminLoginPage/AdminLoginPage';
 import { AdminPage } from 'pages/AdminPage/AdminPage';
 import { LoginPage } from 'pages/LoginPage/LoginPage';
@@ -20,8 +21,8 @@ function App() {
           <Route path="setting" element={<SettingPage />}></Route>
           <Route path="user/self" element={<UserPage />}></Route>
           <Route path="user/self/edit" element={<UserPage />}></Route>
-          <Route path="user/self/follower" element={<UserPage />}></Route>
-          <Route path="user/self/following" element={<UserPage />}></Route>
+          <Route path="user/self/follower" element={<FollowPage />}></Route>
+          <Route path="user/self/following" element={<FollowPage />}></Route>
           <Route path="admin" element={<AdminPage />}></Route>
           <Route path="reply_list" element={<ReplyListPage />}></Route>
           <Route path="reply_list/reply_modal" element={<ReplyListPage />}></Route>
