+++ src/components/FollowList/Follow/Follow.scss
@@ -1,0 +1,66 @@
+.followItem {
+  position: relative;
+  height: auto;
+  margin-top: 20px;
+  border-bottom: 1px solid #E6ECF0;
+  z-index: 0;
+  
+  .name {
+    margin-left: 70px;
+    padding: 10px;
+    color: #171725;
+    font-size: 16px;
+    font-weight: 700;
+  }
+
+  .content {
+    margin-left: 70px;
+    padding: 10px;
+    color: #171725;
+    font-size: 16px;
+    font-weight: 400;
+  }
+
+  .avatarBox {
+    position: absolute;
+    top: 0;
+    left: 20px;
+    width: 50px;
+    height: 50px;
+    cursor: pointer;
+
+    img {
+      width: 100%;
+      height: 100%;
+      border-radius: 50%;
+    }
+  }
+
+  // button styles
+
+  .follow {
+    position: absolute;
+    top: 0;
+    right: 20px;
+    width: 64px;
+    height: 40px;
+    background-color: transparent;
+    border: 1px solid #FF6600;
+    border-radius: 50px;
+    color: #FF6600;
+    cursor: pointer;
+  }
+
+  .following {
+    position: absolute;
+    top: 0;
+    right: 20px;
+    width: 96px;
+    height: 40px;
+    background-color: #FF6600;
+    border: none;
+    border-radius: 50px;
+    color: #FFFFFF;
+    cursor: pointer;
+  }
+}
