+++ src/components/FollowList/FollowList.jsx
@@ -1,4 +1,4 @@
-import './Follow.scss'
+import './FollowList.scss'
 
 // import components
 import { FollowTab } from 'components/Tab/Tab'
@@ -7,11 +7,11 @@ import { Link, useNavigate } from 'react-router-dom'
 // import icons
 import { ReactComponent as BackIcon } from 'assets/icons/back.svg'
 
-export const Follow = () => {
+export const FollowList = () => {
   const navigate = useNavigate()
 
   return (
-    <div className="followContainer">
+    <div className="followListContainer">
       <div className="topSection">
         <div className="title">
           <Link to="/user/self">
@@ -19,7 +19,6 @@ export const Follow = () => {
               <BackIcon />
             </div>
           </Link>
-          
           <div className="titleContent">
             <div className="name">Liz</div>
             <div className="tweets">45 推文</div>
