+++ src/components/FollowList/FollowList.scss
@@ -1,4 +1,4 @@
-.followContainer {
+.followListContainer {
   border: 1px solid #E6ECF0;
   width: 45vw;
   height: 100vh;
