+++ src/components/Tab/Tab.scss
@@ -1,6 +1,7 @@
 .selectionBar {
   display: flex;
-
+  border-bottom: 1px solid #E6ECF0;
+  
   .option {
     display: flex;
     justify-content: center;
