+++ src/pages/FollowPage/FollowPage.jsx
@@ -1,0 +1,14 @@
+import './FollowPage.scss'
+import { SideBar } from "components/SideBar/SideBar";
+import { FollowList } from 'components/FollowList/FollowList';
+import { Recommendation } from 'components/Recommendation/Recommendation';
+
+export const FollowPage = () => {
+  return (
+    <div className="followPageContainer">
+      <SideBar />
+      <FollowList />
+      <Recommendation />
+    </div>
+  )
+}
