+++ src/pages/FollowPage/FollowPage.scss
@@ -1,0 +1,6 @@
+.followPageContainer {
+  position: relative;
+  width: 100vw;
+  display: flex;
+  justify-content: center;  
+}
